Check both Surefire and Failsafe reports

+++ pct.sh
@@ -44,7 +44,7 @@ if grep -q -F -e '<status>INTERNAL_ERROR</status>' pct-report.xml; then
 elif grep -q -F -e '<status>TEST_FAILURES</status>' pct-report.xml; then
 	echo PCT marked failed, checking to see if that is due to a failure to run tests at all
 	for t in pct-work/*/{,*/}target; do
-		if [[ -f "${t}/test-classes/InjectedTest.class" ]] && [[ ! -f "${t}/surefire-reports/TEST-InjectedTest.xml" ]]; then
+		if [[ -f "${t}/test-classes/InjectedTest.class" ]] && [[ ! -f "${t}/surefire-reports/TEST-InjectedTest.xml" ]] && [[ ! -f "${t}/failsafe-reports/TEST-InjectedTest.xml" ]]; then
 			mkdir -p "${t}/surefire-reports"
 			cat >"${t}/surefire-reports/TEST-pct.xml" <<-'EOF'
 				<testsuite name="pct">
